History is now relative to the template schedule
