+++ project/laborganizer/templates/laborganizer/dash-components/history.html
@@ -37,7 +37,7 @@
                 <li class="list-group-item container">
             <div class="row align-items-center">
                 <div class="col-1">
-                <h1 class="ms-2 d-flex justify-content-center">{{node.id}}</h1>
+                <h1 class="ms-2 d-flex justify-content-center">{{node.relative_node_id}}</h1>
                 </div>
                 <div class="col">
                 <!--Change history description-->
