+++ project/quicksched/settings.py
@@ -35,7 +35,7 @@ SECRET_KEY = config('SECRET_KEY')
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = config('DEBUG', default=False, cast=bool)
 
-ALLOWED_HOSTS = []
+ALLOWED_HOSTS = config('ALLOWED_HOSTS', cast=lambda v: [s.strip() for s in v.split(',')])
 
 AUTH_USER_MODEL = "authentication.CustomUserModel"
 AUTHENTICATION_BACKENDS = [
@@ -68,6 +68,10 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
 ]
+
+DJANGO_SUPERUSER_EMAIL=config('DJANGO_SUPERUSER_EMAIL')
+DJANGO_SUPERUSER_PASSWORD=config('DJANGO_SUPERUSER_PASSWORD')
+DJANGO_SUPERUSER_USERNAME=config('DJANGO_SUPERUSER_USERNAME')
 
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
