+++ requirements.txt
@@ -1,0 +1,1 @@
+python-decouple==3.6
